requirements.txt
streamlit==1.28.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.88.0
pandas==1.5.3
numpy==1.21.6
openai==1.0.0
anthropic==0.7.8
google-generativeai==0.3.2
requests==2.31.0
scikit-learn==1.1.3
python-dotenv==1.0.0